Try to pinpoint the intermittent failur of controller_Git.t
This test file randomly crashes after 50 tests with:

Can't call method "request" on an undefined value at /usr/share/perl5/Plack/Test/MockHTTP.pm line 29.

See if the second $site->titles->published_texts->first->full_uri is
the culprit (but that shouldn't return undef) or what.

+++ t/lib/AmuseWiki/Tests.pm
@@ -179,6 +179,7 @@ sub start_jobber {
 
 sub stop_jobber {
     my $pid = shift || $job_pid;
+    die "Missing $pid" unless $pid;
     kill TERM => $pid;
     wait;
     die "$pid is not killed yet" if kill(0, $pid);
